requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
attrs==25.4.0
audioop-lts==0.2.2
cffi==2.0.0
discord.py==2.6.4
frozenlist==1.8.0
idna==3.11
multidict==6.7.0
propcache==0.4.1
pycparser==3.0
PyNaCl==1.6.2
python-dotenv==1.2.1
yarl==1.22.0
yt-dlp==2025.12.8
